Moved world frame conversions to wave_geography
