+++ wave_geography/tests/geography/test_enu_llh_point_conversions.cpp
@@ -35,7 +35,7 @@
 */
 
 #include <gtest/gtest.h>
-#include "wave/utils/world_frame_conversions.hpp"
+#include "wave/geography/world_frame_conversions.hpp"
 
 namespace wave {
